splitted rust into different modules to give a clearer structure
