+++ src-tauri/src/modules/app_state/mod.rs
@@ -1,0 +1,3 @@
+mod commands;
+
+pub use commands::*;
