+++ src-tauri/src/modules/file_system/mod.rs
@@ -1,0 +1,4 @@
+mod helpers;
+mod commands;
+
+pub use commands::*;
