+++ src-tauri/src/modules/mod.rs
@@ -1,0 +1,4 @@
+pub mod types;
+pub mod file_system;
+pub mod profile_manager;
+pub mod app_state;
