+++ src-tauri/src/modules/profile_manager/mod.rs
@@ -1,0 +1,4 @@
+mod helpers;
+mod commands;
+
+pub use commands::*;
