+++ src-tauri/src/modules/types/mod.rs
@@ -1,0 +1,3 @@
+mod tree_data;
+
+pub use tree_data::TreeDataItem;
