+++ src-tauri/src/modules/types/tree_data.rs
@@ -1,0 +1,6 @@
+#[derive(serde::Serialize)]
+pub struct TreeDataItem {
+    pub id: String,
+    pub name: String,
+    pub children: Option<Vec<TreeDataItem>>,
+}
